pyproject.toml
[tool.poetry]
name = "root"
version = "0.1.0"
description = ""
authors = ["Ali Alsarraf"]
license = "MIT"

[tool.poetry.scripts]
manage = "manage:main"

[tool.poetry.dependencies]
python = "3.9.13"
Django = "<4"
python-decouple = "^3.6"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.3"
django-stubs = "^1.12.0"
django-cors-headers = "^3.13.0"
graphene-django = "^2.15.0"
graphene-file-upload = "^1.3.0"
Pillow = "^9.2.0"
django-graphql-auth = "^0.3.16"
django-graphql-jwt = "^0.3.0"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]
annotated-types = "^0.2.0"
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pylint = "^2.14.5"
pylint-django = "^2.5.3"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[tool.mypy]
plugins = "mypy_django_plugin.main"
ignore_missing_imports = true
strict_optional = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
show_error_codes = true



[tool.django-stubs]
django_settings_module = "root.settings"

[tool.pylint.main]
load-plugins = ["pylint_django"]
django-settings-module = "root.settings"

[tool.pylint.message_control]
disable = '''
  attribute-defined-outside-init,
  duplicate-code,
  invalid-name,
  missing-docstring,
  protected-access,
  too-few-public-methods,
  no-member,
  raise-missing-from,
  no-self-argument,
  fixme,
  unsubscriptable-object,
  unused-argument,
  too-many-arguments,
  invalid-str-returned,
  redefined-outer-name,
  inherit-non-class,
  import-outside-toplevel,
  too-many-ancestors,
  cyclic-import,
  format,
  wrong-import-order,
  ungrouped-imports,
  wrong-import-position
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "root.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = "--cov=. --no-cov-on-fail --cov-report term-missing:skip-covered"
filterwarnings = [
  "ignore::django.utils.deprecation.RemovedInDjango40Warning",
  "ignore::django.utils.deprecation.RemovedInDjango41Warning",
]

[tool.coverage.run]
omit = []
